requirements.txt
feedparser
python-vlc
